+++ src/apis/employee/mock/getMyApplication.mock.ts
@@ -1,6 +1,6 @@
 import { http, HttpResponse } from 'msw';
 import { getMyApplicationPath } from '../hooks/useGetMyApplication';
-import { myRecruitList } from '@/pages/employee/myPage/data/index.mock';
+import { myRecruitList } from '@/pages/myPage/employee/data/index.mock';
 
 export const EmployeePageMockHandler = [
   http.get(getMyApplicationPath(), () => {
