+++ src/routes/router.tsx
@@ -6,14 +6,14 @@ import LoadingPage from '@/pages/auth/Loading';
 import App from '@/App';
 import Recruit from '@/pages/recruit';
 import RegisterVisa from '@/pages/registerVisa/RegisterVisa';
-import PostNotice from '@/pages/employer/postNotice/PostNotice';
+import PostNotice from '@/pages/postNotice/PostNotice';
 import Home from '@/pages/home';
 import ApplyGuide from '@/pages/apply/applyguide/ApplyGuide';
 import ApplyPage from '@/pages/apply/applypage/ApplyPage';
 import MyCompany from '@/pages/myCompany/MyCompany';
 import Applicants from '@/pages/applicants/Applicants';
 import Resume from '@/pages/resume/Resume';
-import EmployeeMyPage from '@/pages/employee/myPage/EmployeeMyPage';
+import EmployeeMyPage from '@/pages/myPage/employee/EmployeeMyPage';
 import RegisterSign from '@/pages/registerSign/RegisterSign';
 import RegisterCompany from '@/pages/registerCompany/RegisterCompany';
 import EmployerMyPage from '@/pages/myPage/employer/EmPloyerMyPage';
